automatically use retarget macro to fix target frame in ui
There is no longer a need for putting a /cleartarget /targetlasttarget
macro at the top of the combat or healing procedures. Just make sure
you have a macro as such:

/cleartarget
/targetlasttarget

Name it anything you like. It will automatically be used any time
your target is changed in memory, thus fixing the glitch with the lua
target frame.